refactor: dex support that does not support deadline
