+++ contracts/terraswap_pair/src/testing.rs
@@ -1,6 +1,6 @@
 use crate::contract::{
-    assert_deadline, assert_max_spread, assert_minimum_assets, execute, instantiate,
-    query_pair_info, query_pool, query_reverse_simulation, query_simulation, reply,
+    assert_max_spread, assert_minimum_assets, execute, instantiate, query_pair_info, query_pool,
+    query_reverse_simulation, query_simulation, reply,
 };
 use crate::error::ContractError;
 use classic_terraswap::mock_querier::mock_dependencies;
@@ -577,7 +577,10 @@ fn withdraw_liquidity() {
     let env = mock_env();
     let info = mock_info("liquidity0000", &[]);
     let err = execute(deps.as_mut(), env, info, msg).unwrap_err();
-    assert_eq!(err, ContractError::ExpiredDeadline {})
+    assert_eq!(
+        err,
+        ContractError::Std(StdError::generic_err("Expired deadline"))
+    )
 }
 
 #[test]
@@ -1547,26 +1550,4 @@ fn test_assert_minimum_assets_with_unknown_asset() {
             asset: "0ukrw".to_string()
         }
     )
-}
-
-#[test]
-fn test_assert_deadline_with_normal() {
-    assert_deadline(5u64, Some(10u64)).unwrap();
-}
-
-#[test]
-fn test_assert_deadline_with_expired() {
-    let err = assert_deadline(10u64, Some(5u64)).unwrap_err();
-    assert_eq!(err, ContractError::ExpiredDeadline {})
-}
-
-#[test]
-fn test_assert_deadline_with_same() {
-    let err = assert_deadline(10u64, Some(10u64)).unwrap_err();
-    assert_eq!(err, ContractError::ExpiredDeadline {})
-}
-
-#[test]
-fn test_assert_deadline_with_none() {
-    assert_deadline(5u64, None).unwrap();
 }
