+++ contracts/terraswap_router/src/operations.rs
@@ -12,6 +12,7 @@ use classic_terraswap::asset::{Asset, AssetInfo, PairInfo};
 use classic_terraswap::pair::ExecuteMsg as PairExecuteMsg;
 use classic_terraswap::querier::{query_balance, query_pair_info, query_token_balance};
 use classic_terraswap::router::SwapOperation;
+use classic_terraswap::util::assert_deadline;
 use cw20::Cw20ExecuteMsg;
 use terra_cosmwasm::{create_swap_msg, create_swap_send_msg, TerraMsgWrapper};
 
@@ -28,6 +29,8 @@ pub fn execute_swap_operation(
     if env.contract.address != info.sender {
         return Err(StdError::generic_err("unauthorized"));
     }
+
+    assert_deadline(env.block.time.seconds(), deadline)?;
 
     let messages: Vec<CosmosMsg<TerraMsgWrapper>> = match operation {
         SwapOperation::NativeSwap {
@@ -92,7 +95,6 @@ pub fn execute_swap_operation(
                 offer_asset,
                 None,
                 to,
-                deadline,
             )?]
         }
         SwapOperation::Loop {
@@ -128,7 +130,6 @@ pub fn execute_swap_operation(
                 offer_asset,
                 None,
                 to,
-                deadline,
             )?]
         }
         SwapOperation::Astroport {
@@ -164,7 +165,6 @@ pub fn execute_swap_operation(
                 offer_asset,
                 Some(Decimal::from_str("0.5")?),
                 to,
-                deadline,
             )?]
         }
     };
@@ -178,7 +178,6 @@ pub fn asset_into_swap_msg(
     offer_asset: Asset,
     max_spread: Option<Decimal>,
     to: Option<String>,
-    deadline: Option<u64>,
 ) -> StdResult<CosmosMsg<TerraMsgWrapper>> {
     match offer_asset.info.clone() {
         AssetInfo::NativeToken { denom } => {
@@ -200,7 +199,7 @@ pub fn asset_into_swap_msg(
                     belief_price: None,
                     max_spread,
                     to,
-                    deadline,
+                    deadline: None,
                 })?,
             }))
         }
@@ -215,7 +214,7 @@ pub fn asset_into_swap_msg(
                     belief_price: None,
                     max_spread,
                     to,
-                    deadline,
+                    deadline: None,
                 })?,
             })?,
         })),
