+++ packages/classic_terraswap/src/lib.rs
@@ -4,6 +4,7 @@ pub mod pair;
 pub mod querier;
 pub mod router;
 pub mod token;
+pub mod util;
 
 #[cfg(not(target_arch = "wasm32"))]
 pub mod mock_querier;
